First working version
Support for mp3 files and azlyrics as a source, methods implemented for parsing automatically and semi-manually through a folder's files to find and embed lyrics to compatible files.
The folder parsing and lyrics fetching methods in the test class can result in an ip ban if not throttled, currently using a semaphore with thread awaits of 3-4 seconds between html requests to avoid being bot flagged.